backend/requirements.txt
Flask
mysql-connector-python
python-dotenv
gunicorn
clerk-sdk
pyjwt